+++ typo3scan.py
@@ -18,7 +18,7 @@
 # along with this program. If not, see [http://www.gnu.org/licenses/](http://www.gnu.org/licenses/)
 #-------------------------------------------------------------------------------
 
-__version__ = '1.0.2'
+__version__ = '1.1'
 __program__ = 'Typo3Scan'
 __description__ = 'Automatic Typo3 enumeration tool'
 __author__ = 'https://github.com/whoot'
@@ -170,7 +170,6 @@ Options:
 
   General:
     -u | --update       Update extensions and vulnerability database.
-    -r | --reset        Reset the database.
     --core VERSION      Show all known vulnerabilities for given Typo3 version.
     --ext EXT:VERSION   Show all known vulnerabilities for given extension and version.
 """)
@@ -181,7 +180,6 @@ Options:
     group.add_argument('-f', '--file', dest='file')
     group.add_argument('-d', '--domain', dest='domain', type=str, nargs='+')
     group.add_argument('-u', '--update', dest='update', action='store_true')
-    group.add_argument('-r', '--reset', dest='reset', action='store_true')
     group.add_argument('--core', dest='core', type=str)
     group.add_argument('--ext', dest='extension', type=str)
     parser.add_argument('--force', dest='force', action='store_true')
@@ -200,10 +198,6 @@ Options:
     if args.help or not len(sys.argv) > 1:
         print_help()
 
-    elif args.reset:
-        from lib.initdb import DB_Init
-        DB_Init()
-
     elif args.update:
         from lib.update import Update
         Update()
@@ -212,13 +206,13 @@ Options:
         database = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'lib', 'typo3scan.db')
         conn = sqlite3.connect('lib/typo3scan.db')
         c = conn.cursor()
-        c.execute('SELECT advisory, vulnerability, subcomponent, affected_version_max, affected_version_min FROM core_vulns WHERE (?<=affected_version_max AND ?>=affected_version_min)', (args.core, args.core,)) 
+        c.execute('SELECT advisory, vulnerability, subcomponent, affected_version_max, affected_version_min, severity FROM core_vulns WHERE (?<=affected_version_max AND ?>=affected_version_min)', (args.core, args.core,)) 
         data = c.fetchall()
         json_list = {}
         if data:
             for vulnerability in data:
                 if parse_version(args.core) <= parse_version(vulnerability[3]):
-                    json_list[vulnerability[0]] = {'Type': vulnerability[1], 'Subcomponent': vulnerability[2], 'Affected': '{} - {}'.format(vulnerability[3], vulnerability[4]), 'Advisory': 'https://typo3.org/security/advisory/{}'.format(vulnerability[0].lower())}
+                    json_list[vulnerability[0]] = {'Type': vulnerability[1], 'Subcomponent': vulnerability[2], 'Affected': '{} - {}'.format(vulnerability[3], vulnerability[4]), 'Severity': vulnerability[5],'Advisory': 'https://typo3.org/security/advisory/{}'.format(vulnerability[0].lower())}
             if json_list:
                 print(Style.BRIGHT + '\n[+] Known Vulnerabilities for Typo3 v{}\n'.format(args.core) + Style.RESET_ALL)
                 for vulnerability in json_list.keys():
@@ -226,6 +220,7 @@ Options:
                     print('    \u251c Vulnerability Type:'.ljust(28) + json_list[vulnerability]['Type'])
                     print('    \u251c Subcomponent:'.ljust(28) + json_list[vulnerability]['Subcomponent'])
                     print('    \u251c Affected Versions:'.ljust(28) + json_list[vulnerability]['Affected'])
+                    print('    \u251c Severity:'.ljust(28) + json_list[vulnerability]['Severity'])
                     print('    \u2514 Advisory URL:'.ljust(28) + json_list[vulnerability]['Advisory'] + '\n')
         if not json_list:
             print('\n' + Fore.GREEN + Style.BRIGHT + '[+] Typo3 v{} has no known vulnerabilities\n'.format(args.core) + Style.RESET_ALL)
@@ -249,13 +244,13 @@ Options:
             print('\n' + Fore.RED + Style.BRIGHT + '[!] Extension \'{}\' does not exist\n'.format(name) + Style.RESET_ALL)
             sys.exit(-1)
         else:
-            c.execute('SELECT advisory, vulnerability, affected_version_max, affected_version_min FROM extension_vulns WHERE (extensionkey=? AND ?<=affected_version_max AND ?>=affected_version_min)', (name, version, version,))
+            c.execute('SELECT advisory, vulnerability, affected_version_max, affected_version_min, severity FROM extension_vulns WHERE (extensionkey=? AND ?<=affected_version_max AND ?>=affected_version_min)', (name, version, version,))
             data = c.fetchall()
         json_list = {}
         if data:
             for vulnerability in data:
                 if parse_version(version) <= parse_version(vulnerability[2]):
-                    json_list[vulnerability[0]] = {'Type': vulnerability[1], 'Affected': '{} - {}'.format(vulnerability[2], vulnerability[3]), 'Advisory': 'https://typo3.org/security/advisory/{}'.format(vulnerability[0].lower())}
+                    json_list[vulnerability[0]] = {'Type': vulnerability[1], 'Affected': '{} - {}'.format(vulnerability[2], vulnerability[3]), 'Severity': vulnerability[4], 'Advisory': 'https://typo3.org/security/advisory/{}'.format(vulnerability[0].lower())}
             if json_list:
                 if version == '0.0.0':
                     print(Style.BRIGHT + '\n[+] Known vulnerabilities for \'{}\'\n'.format(name) + Style.RESET_ALL)
@@ -265,6 +260,7 @@ Options:
                     print(Style.BRIGHT + '   [!] {}'.format(Fore.RED + vulnerability + Style.RESET_ALL))
                     print('    \u251c Vulnerability Type: '.ljust(28) + json_list[vulnerability]['Type'])
                     print('    \u251c Affected Versions: '.ljust(28) + '{}'.format(json_list[vulnerability]['Affected']))
+                    print('    \u251c Severity:'.ljust(28) + json_list[vulnerability]['Severity'])
                     print('    \u2514 Advisory URL:'.ljust(28) + '{}\n'.format(json_list[vulnerability]['Advisory'].lower()))
         if not json_list:
             print('\n' + Fore.GREEN + Style.BRIGHT + '[+] \'{}\' v{} has no known vulnerabilities\n'.format(name, version) + Style.RESET_ALL)